requirements.txt
absl-py==2.1.0
accelerate==1.9.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
apturl==0.5.2
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
boto3==1.35.13
botocore==1.35.13
Brlapi==0.8.3
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
clarabel==0.10.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cupshelpers==1.0
cvxopt==1.3.2
cvxpy==1.6.0
cycler==0.12.1
dask==2024.9.0
dask-expr==1.1.14
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dm-tree==0.1.9
docopt==0.6.2
duplicity==0.8.21
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fancyimpute==0.7.0
fasteners==0.14.1
fastjsonschema==2.20.0
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
future==0.18.2
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.66.1
h2o==3.46.0.3
h5py==3.11.0
hf-xet==1.1.7
httplib2==0.20.2
huggingface-hub==0.34.3
idna==3.3
imbalanced-learn==0.12.0
imblearn==0.0
import-ipynb==0.1.4
importlib_metadata==8.5.0
iniconfig==2.0.0
ipython==8.12.3
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.9.2
keras-core==0.1.7
keras-tuner==1.4.7
keras-tuner-cv==1.0.3
keyring==23.5.0
kiwisolver==1.4.7
knnimpute==0.1.0
kt-legacy==1.0.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lightgbm==4.5.0
lightning-utilities==0.11.7
llvmlite==0.43.0
locket==1.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
modin==0.32.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.11.0
networkx==3.3
nose==1.3.7
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
osqp==0.6.7.post3
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.4
partd==1.4.2
patsy==0.5.6
peft==0.17.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.5.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
pydicom==2.4.4
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyspark==3.5.1
pytest==8.2.2
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytorch-lightning==2.3.0
pytorch-pretrained-bert==0.6.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.2.0
qdldl==0.1.7.post5
ray==2.37.0
referencing==0.35.1
regex==2024.7.24
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.6.1
scikit-learn==1.5.1
scipy==1.13.1
screen-resolution-extra==0.0.0
scs==3.2.7.post2
seaborn==0.13.2
SecretStorage==3.3.1
shap==0.45.1
six==1.16.0
slicer==0.0.8
smmap==5.0.1
soupsieve==2.6
SQLAlchemy==2.0.19
ssh-import-id==5.11
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.14.0
systemd-python==234
tabulate==0.9.0
te2rules==1.0.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.21.4
tomli==2.0.1
toolz==0.12.1
torch==2.7.1
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.55.0
triton==3.3.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.14.1
xdg==5
xgboost==2.0.3
xkit==0.0.0
yarg==0.1.9
yarl==1.9.11
zipp==3.20.2
